backend/requirements.txt
Django
djangorestframework
django-crispy-forms
django-filter
django-request
django-guardian
Markdown
Pigments
coreapi
psycopg2
requests
lxml
bs4
